Initialize "Boostan" Django project
The initial django files adds by `django-admin.py startproject boostan`
command.
The "Boostan" project is an opensource and free alternative for
"Golestan EMS" that planned to be written in python/django.
The "Golestan EMS" itself said to be written in ASP.NET and as not a
free web-developement language and poor preformance decided to make
an alternative for this program that is free and opensource and written
in python/django so it's a preformance-friendly project and avoided
a lots of bugs and (security) problems initially.

In The Hope of be a usefull project and helpfull for having a better
world of free software!

End of Initial Commit Message!